requirements.txt
mcp 
mcp[cli]
python-dotenv
openai
langchain
langchain-openai
langchain-mcp-adapters
langgraph